exynos9820-common: Switch to Vulkan UI renderer
We tried doing this on Android 11 and later reverted it due to fd leaks
during prolonged video calls. However, the issue has been fixed on
Android 12, so let's enable it again for performance.

Change-Id: I1cd2ca6931aad60f9c48d40aa0212b69c68edef8

exynos9810-common: Switch to Vulkan UI renderer
We tried doing this on Android 11 and later reverted it due to fd leaks
during prolonged video calls. However, the issue has been fixed on
Android 12, so let's enable it again for performance.

Change-Id: I1cd2ca6931aad60f9c48d40aa0212b69c68edef8